Adding Twitter Share GitHub Action
This action will generate text for the user to copy paste and share on Twitter, including:
 - a custom message
 - hashtags
 - "at" usernames

And specifically, it includes links to one or more files that match a pattern from the pull request just merged. I found I needed to do this because the traditional Twitter share buttons didn't work (we can't embed the widget.js from Twitter in GitHub) and then including the hashtags and "at" references in a URL to open the same interface would be truncated in the URL. The use case for this action is any GitHub user/organization that would want to encourage users to contribute content, and then be given a fun Tweet to share that.